requirements.txt
alpaca-py==0.16.0
flask==3.0.2
pandas==2.2.1
requests==2.28.1
json==2.0.9
